requirements.txt
httplib2
google-api-python-client
beautifulsoup4
python-dateutil
pytest